Add Python implementations of certain longobject.c functions. These use asymptotically faster algorithms that can be
used for operations on integers with many digits.  In those cases, the performance overhead of the Python implementation is not significant since the asymptotic behavior is what dominates runtime. Functions provided by this module should be considered private and not part of any public API.

Currently implemented is `int_to_decimal_string()` (based on code from Tim), `int_divmod()` (based on code from Mark), and `str_to_int()` (based on code from bjorn-martinsson).  The longobject.c module has been changed to call into `_pylong` if the Python version is likely to be faster.


Examples of performance improvements:

int->str
```
python -m timeit -s "i=(1<<2_000_000) - 1" "str(i)" 
Before: 1 loop, best of 5: 3.58 sec per loop
After: 2 loops, best of 5: 168 msec per loop
```
str->int
```
python -m timeit -s "s='1'*1_000_000" "int(s)"
Before: 1 loop, best of 5: 3.65 sec per loop
After: 1 loop, best of 5: 470 msec per loop
```
divmod
```
python -m timeit -s "a=(1<<20_000_000)-1; b=(1<<100_000)" "divmod(a, b)"
Before: 1 loop, best of 5: 2.23 sec per loop
After: 2 loops, best of 5: 141 msec per loop
```


Co-author: Tim Peters <tim.peters@gmail.com>
Co-author: Mark Dickinson <dickinsm@gmail.com>
Co-author: Bjorn Martinsson

<!-- gh-issue-number: gh-90716 -->
* Issue: gh-90716
<!-- /gh-issue-number -->