@nascheme 2022-09-08
Some relevant discussion on the pros and cons of adding these "smarter" but more complicated algorithms:

https://discuss.python.org/t/faster-large-integer-multiplication/13300/26

I think implementing them in Python lowers the bar quite a bit in terms of what is maintainable in the long term.  We have a small number of core devs who are numeric experts so we need to be cautious about adding a bunch of complex code.  Maybe this PR already goes too far?  It feels like pretty good "bang for buck" to me.