@bjorn-martinsson 2022-09-08
About long_to_decimal, here are some things that I've noticed/thoughts.

1. Increasing the size of BITLIM to something like 32768 gives a noticable performance boost for me.
2. I think it is more natural to add this algorithm to the decimal module. This code is essentially just a smart way of implementing the constructor of `decimal.Decimal` when called with an int. 
3. The code is really slow in PyPy. I'm guessing this is because decimal is slow in PyPy. 
4. I think the current implementation is a bit hard to understand. Maybe something like this would make it easier to follow:

```py
D = decimal.Decimal
mem = {1: D(2)}
def w2pow(w):
    """ return D(2)**w and store the result """
    if w not in mem:
        if w - 1 in mem:
            mem[w] = mem[w - 1] + mem[w - 1]
        else:
            w2 = w >> 1
            mem[w] = w2pow(w2) * w2pow(w - w2)
    return mem[w]

BITLIM = 32768
def inner(n, w):
    if w <= BITLIM:
        return D(n)
    w2 = w >> 1
    hi = n >> w2
    lo = n - (hi << w2)
    return inner(hi, w - w2) * w2pow(w2) + inner(lo, w2)

with decimal.localcontext() as ctx:
    ctx.prec = decimal.MAX_PREC
    ctx.Emax = decimal.MAX_EMAX
    ctx.Emin = decimal.MIN_EMIN
    ctx.traps[decimal.Inexact] = 1
    
    result = inner(n, n.bit_length())
``` 

Btw there is an `if 0:` in `divmod_fast` which I'm assuming is not supposed to be there.