@bjorn-martinsson 2022-09-09
I agree that persistent dictionaries are bad, but I never intended mem to be one.

About the number of keys used. I fixed my implementation to use fewer keys. Now I believe it uses the same number of keys, or sometimes 1 less key than your implementation. For example, try it out with bit width 1,000,001. For `BITLIM=128` I use 15 keys, and your implementation uses 16.

```py
mem = {}
def w2pow(w):
    """ return D(2)**w and store the result """
    if w not in mem:
        if w - 1 in mem:
            mem[w] = mem[w - 1] + mem[w - 1]
        elif w <= BITLIM:
            mem[w] = D2 ** w
        else:
            w2 = w >> 1
            mem[w] = w2pow(w2) * w2pow(w - w2)
    return mem[w]

BITLIM = 32768 # 128
def inner(n, w):
    if w <= BITLIM:
        return D(n)
    w2 = w >> 1
    hi = n >> w2
    lo = n - (hi << w2)
    return inner(lo, w2) + inner(hi, w - w2) * w2pow(w2)

with decimal.localcontext() as ctx:
    ctx.prec = decimal.MAX_PREC
    ctx.Emax = decimal.MAX_EMAX
    ctx.Emin = decimal.MIN_EMIN
    ctx.traps[decimal.Inexact] = 1
    
    result = inner(n, n.bit_length())
```
I don't think it is bad to generate keys with recursion, since `inner` already uses the same recursion. Also, note that only the initial call to `w2pow()` will recurse deep, every call after that will not do any recursion at all.