@tim-one 2022-09-09
I'm not going to argue about this. As I said, I didn't find the code hard to understand to begin with, so this is just random thrashing to me. I expect you find your code easier to understand because you wrote it.

> With this implementation, `w2pow()` is called in smart order (it is called with small values of `w` first, then bigger values of w), making it only have recursion depth at most 2.

That seems wrong.  Trace through, e.g., `w2pow(2**20)`. That invokes

```py
mem[w] = w2pow(w2) * w2pow(w - w2)
```

which makes a recursive call with `2**19`, which in turn makes a recursive call on the same line with `2**18`, and so on, until the function is finally called with `2**15` (your `BITLIM` value). Only then does the recursion start to unwind.