@nascheme 2022-09-09
Hi [bjorn-martinsson](https://github.com/bjorn-martinsson), I appreciate the interest in improving the functions in `_pylong.py` but polishing the functions is not the primary focus of this PR.  Perhaps we should create another issue to focus on that?  If they produce correct results and perform reasonably well, I think that's good enough for now.  So, I feel like `long_to_decimal()` is in pretty okay shape already.

I'm less confident about `divmod_fast()` since Mark has not looked at this PR yet, AFAIK.  I extracted the code from an old issue and I believe it was some prototype code, not intended for serious use.

Writing more extensive unit tests and scripts to produce timing/benchmark reports is perhaps the next step.  `long_to_decimal` doesn't handle the bytes writer case, need to fix that.  Need to come up with better threshold tests of when to switch to the _pylong functions.  I'm using `Py_ABS(Py_SIZE(n)) > 1000` but that's almost certainly not the right thing.


Based on [comments from njs](https://discuss.python.org/t/int-str-conversions-broken-in-latest-python-bugfix-releases/18889/8), I wonder if _pylong should use ContextVar and then we could move the `max_str_digits` checking into there.  Getting away from a global interpreter setting would be quite a nice improvement.  Doing it inside `_pylong` should be a lot easier.