@tim-one 2022-09-09
Neil, I think the one _basic_ thing this is missing is an asymptotically faster str->int function. Conversions in both directions were the real topic of the original issue report, although it focused on int->str (as "the harder" problem).  Both are quadratic-time today.

@bjorn-martinsson  posted suitable code for str->int (in a comment) in gh-90716.  It's much like the int->str here, but sticks to native Python int arithmetic so only gets "Karatsuba-like" speedup. Which is huge, but not _as_ huge as we can get from `decimal`'s fancier `*`. But I don't see a sane way to apply `decimal` in the decimal string -> binary int direction. The result has to be a CPython bigint, and conversion from `Decimal` to `int` is also quadratic-time(*).

(*) But could be faster after _this_ PR lands (convert the `Decimal` to a decimal string (linear time in # of digits), then apply the hypothetical new faster-than-quadratic str->int function).